+++ packages/cli/plugin-bff/src/cli.ts
@@ -4,13 +4,127 @@ import { ApiRouter } from '@modern-js/bff-core';
 import { compile } from '@modern-js/server-utils';
 import type { ServerRoute } from '@modern-js/types';
 import { fs, API_DIR, SHARED_DIR, normalizeOutputPath } from '@modern-js/utils';
+import clientGenerator from './utils/clientGenerator';
+import pluginGenerator from './utils/pluginGenerator';
+import runtimeGenerator from './utils/runtimeGenerator';
 
 const DEFAULT_API_PREFIX = '/api';
 const TS_CONFIG_FILENAME = 'tsconfig.json';
+const RUNTIME_CREATE_REQUEST = '@modern-js/plugin-bff/runtime/create-request';
 
 export const bffPlugin = (): CliPlugin<AppTools> => ({
   name: '@modern-js/plugin-bff',
   setup: api => {
+    const compileApi = async () => {
+      const {
+        appDirectory,
+        distDirectory,
+        apiDirectory,
+        sharedDirectory,
+        moduleType,
+      } = api.useAppContext();
+      const modernConfig = api.useResolvedConfigContext();
+
+      const distDir = path.resolve(distDirectory);
+      const apiDir = apiDirectory || path.resolve(appDirectory, API_DIR);
+      const sharedDir =
+        sharedDirectory || path.resolve(appDirectory, SHARED_DIR);
+      const tsconfigPath = path.resolve(appDirectory, TS_CONFIG_FILENAME);
+
+      const sourceDirs = [];
+      if (await fs.pathExists(apiDir)) {
+        sourceDirs.push(apiDir);
+      }
+
+      if (await fs.pathExists(sharedDir)) {
+        sourceDirs.push(sharedDir);
+      }
+
+      const { server } = modernConfig;
+      const { alias } = modernConfig.source;
+      const { babel } = modernConfig.tools;
+
+      if (sourceDirs.length > 0) {
+        await compile(
+          appDirectory,
+          {
+            server,
+            alias,
+            babelConfig: babel,
+          },
+          {
+            sourceDirs,
+            distDir,
+            tsconfigPath,
+            moduleType,
+          },
+        );
+      }
+    };
+
+    const generator = async () => {
+      const { appDirectory, apiDirectory, lambdaDirectory, port } =
+        api.useAppContext();
+
+      const modernConfig = api.useResolvedConfigContext();
+      const relativeDistPath = modernConfig?.output?.distPath?.root || 'dist';
+      const { bff } = modernConfig || {};
+      const prefix = bff?.prefix || DEFAULT_API_PREFIX;
+      const httpMethodDecider = bff?.httpMethodDecider;
+
+      const apiRouter = new ApiRouter({
+        apiDir: apiDirectory,
+        appDir: appDirectory,
+        lambdaDir: lambdaDirectory,
+        prefix,
+        httpMethodDecider,
+        isBuild: true,
+      });
+
+      const lambdaDir = apiRouter.getLambdaDir();
+      const existLambda = apiRouter.isExistLambda();
+
+      const runtime =
+        (bff as any)?.runtimeCreateRequest || RUNTIME_CREATE_REQUEST;
+      const relativeApiPath = path.relative(appDirectory, apiDirectory);
+      const relativeLambdaPath = path.relative(appDirectory, lambdaDir);
+
+      await pluginGenerator({
+        prefix,
+        appDirectory,
+        relativeDistPath,
+        relativeApiPath,
+        relativeLambdaPath,
+      });
+      await clientGenerator({
+        prefix,
+        appDir: appDirectory,
+        apiDir: apiDirectory,
+        lambdaDir,
+        existLambda,
+        port,
+        requestCreator: (bff as any)?.requestCreator,
+        httpMethodDecider,
+        relativeDistPath,
+        relativeApiPath,
+      });
+      await runtimeGenerator({
+        runtime,
+        appDirectory,
+        relativeDistPath,
+      });
+    };
+
+    const handleCrossProjectInvocation = async (isBuild = false) => {
+      const { bff } = api.useResolvedConfigContext();
+      if (bff?.enableCrossProjectInvocation) {
+        if (!isBuild) {
+          await compileApi();
+        }
+        await generator();
+      }
+    };
+
     return {
       config() {
         return {
@@ -60,11 +174,10 @@ export const bffPlugin = (): CliPlugin<AppTools> => ({
                 });
 
               chain.resolve.alias.set('@api', apiDirectory);
+
               chain.resolve.alias.set(
                 '@modern-js/runtime/bff',
-                isServer
-                  ? require.resolve('@modern-js/create-request/server')
-                  : require.resolve('@modern-js/create-request/client'),
+                RUNTIME_CREATE_REQUEST,
               );
             },
           },
@@ -117,51 +230,39 @@ export const bffPlugin = (): CliPlugin<AppTools> => ({
 
         return { plugins };
       },
+      async beforeDev() {
+        await handleCrossProjectInvocation();
+      },
 
       async afterBuild() {
-        const {
-          appDirectory,
-          distDirectory,
-          apiDirectory,
-          sharedDirectory,
-          moduleType,
-        } = api.useAppContext();
-        const modernConfig = api.useResolvedConfigContext();
+        await compileApi();
+        await handleCrossProjectInvocation(true);
+      },
+      async watchFiles() {
+        const appContext = api.useAppContext();
+        const config = api.useResolvedConfigContext();
 
-        const distDir = path.resolve(distDirectory);
-        const apiDir = apiDirectory || path.resolve(appDirectory, API_DIR);
-        const sharedDir =
-          sharedDirectory || path.resolve(appDirectory, SHARED_DIR);
-        const tsconfigPath = path.resolve(appDirectory, TS_CONFIG_FILENAME);
-
-        const sourceDirs = [];
-        if (await fs.pathExists(apiDir)) {
-          sourceDirs.push(apiDir);
+        if (config?.bff?.enableCrossProjectInvocation) {
+          return [appContext.apiDirectory];
+        } else {
+          return [];
         }
+      },
 
-        if (await fs.pathExists(sharedDir)) {
-          sourceDirs.push(sharedDir);
-        }
-
-        const { server } = modernConfig;
-        const { alias } = modernConfig.source;
-        const { babel } = modernConfig.tools;
-
-        if (sourceDirs.length > 0) {
-          await compile(
-            appDirectory,
-            {
-              server,
-              alias,
-              babelConfig: babel,
-            },
-            {
-              sourceDirs,
-              distDir,
-              tsconfigPath,
-              moduleType,
-            },
-          );
+      async fileChange(e: {
+        filename: string;
+        eventType: string;
+        isPrivate: boolean;
+      }) {
+        const { filename, eventType, isPrivate } = e;
+        const { appDirectory, apiDirectory } = api.useAppContext();
+        const relativeApiPath = path.relative(appDirectory, apiDirectory);
+        if (
+          !isPrivate &&
+          (eventType === 'change' || eventType === 'unlink') &&
+          filename.startsWith(`${relativeApiPath}/`)
+        ) {
+          await handleCrossProjectInvocation();
         }
       },
     };