+++ packages/runtime/plugin-garfish/src/cli/hooks.ts
@@ -1,0 +1,6 @@
+import type { Entrypoint } from '@modern-js/types';
+
+export type AppendEntryCodeFn = (params: {
+  entrypoint: Entrypoint;
+  code: string;
+}) => string | Promise<string>;