+++ packages/runtime/plugin-router-v5/src/runtime/hooks.ts
@@ -1,6 +1,9 @@
-import { createWaterfall } from '@modern-js/plugin';
+import { createSyncHook } from '@modern-js/plugin-v2';
 import type { RouteProps } from 'react-router-dom';
+import type { SingleRouteConfig } from './plugin';
 
-const modifyRoutesHook = createWaterfall<RouteProps[]>();
+// only for inhouse use
+const modifyRoutesHook =
+  createSyncHook<(routes: RouteProps[]) => SingleRouteConfig[]>();
 
 export { modifyRoutesHook };