+++ packages/runtime/plugin-runtime/src/router/runtime/types.ts
@@ -39,6 +39,10 @@ export type RouterConfig = {
   future?: Partial<{
     v7_startTransition: boolean;
   }>;
+  /**
+   * An unstable feature, which will reload the page when the current browser URL and the SSR Context URL do not match.
+   */
+  unstable_reloadOnURLMismatch?: boolean;
 };
 
 export type Routes = RouterConfig['routesConfig']['routes'];