+++ packages/toolkit/plugin-v2/src/cli/api.ts
@@ -5,9 +5,11 @@ import type { AppContext, InternalContext } from '../types/cli/context';
 import type { CLIPluginExtends } from '../types/cli/plugin';
 import type { PluginManager } from '../types/plugin';
 import type { DeepPartial } from '../types/utils';
+import { debug } from './run/utils/debug';
 
 export function initPluginAPI<Extends extends CLIPluginExtends>({
   context,
+  pluginManager,
 }: {
   context: InternalContext<Extends>;
   pluginManager: PluginManager;
@@ -73,7 +75,8 @@ export function initPluginAPI<Extends extends CLIPluginExtends>({
     context = merge(context, updateContext);
   }
 
-  return {
+  const pluginAPI = {
+    isPluginExists: pluginManager.isPluginExists,
     getAppContext,
     getConfig,
     getNormalizedConfig,
@@ -99,6 +102,7 @@ export function initPluginAPI<Extends extends CLIPluginExtends>({
     onFileChanged: hooks.onFileChanged.tap,
     onBeforeRestart: hooks.onBeforeRestart.tap,
     onBeforeCreateCompiler: hooks.onBeforeCreateCompiler.tap,
+    onDevCompileDone: hooks.onDevCompileDone.tap,
     onAfterCreateCompiler: hooks.onAfterCreateCompiler.tap,
     onBeforeBuild: hooks.onBeforeBuild.tap,
     onAfterBuild: hooks.onAfterBuild.tap,
@@ -109,4 +113,19 @@ export function initPluginAPI<Extends extends CLIPluginExtends>({
     onBeforeExit: hooks.onBeforeExit.tap,
     ...extendsPluginApi,
   };
+
+  return new Proxy(pluginAPI, {
+    get(target: Record<string, any>, prop: string) {
+      // hack then function to fix p-defer handle error
+      if (prop === 'then') {
+        return undefined;
+      }
+      if (prop in target) {
+        return target[prop];
+      }
+      return () => {
+        debug(`api.${prop.toString()} not exist`);
+      };
+    },
+  }) as CLIPluginAPI<Extends>;
 }