+++ packages/document/module-doc/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @modern-js/module-tools-docs
 
+## 2.63.7
+
 ## 2.63.6
 
 ## 2.63.5