+++ tests/e2e/builder/cases/import-antd-v5/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @e2e/webpack-builder-import-antd-v5
 
+## 2.63.7
+
 ## 2.63.6
 
 ## 2.63.5