+++ tests/integration/load-config/fixtures/async-config-function/CHANGELOG.md
@@ -1,5 +1,7 @@
 # integration-config-async-config-test
 
+## 2.63.7
+
 ## 2.63.6
 
 ## 2.63.5